move Checker to coprocessor
