+++ src/raftstore/store/mod.rs
@@ -41,4 +41,3 @@ pub use self::peer_storage::{do_snapshot, CacheQueryStats, PeerStorage, SnapStat
                              RAFT_INIT_LOG_INDEX, RAFT_INIT_LOG_TERM};
 pub use self::snap::{check_abort, copy_snapshot, ApplyOptions, SnapEntry, SnapKey, SnapManager,
                      Snapshot, SnapshotDeleter, SnapshotStatistics};
-pub use self::worker::SplitChecker;
